# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PFS Obslog API",
    "description": "PFS Observation Log System API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/pfs_designs": {
      "get": {
        "tags": [
          "pfs_designs",
          "pfs_designs"
        ],
        "summary": "List PFS Designs",
        "description": "Get a paginated list of PFS Design files with optional filtering and sorting.",
        "operationId": "list_pfs_designs_api_pfs_designs_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search string (matches name or id)",
              "title": "Search"
            },
            "description": "Search string (matches name or id)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "date_modified",
                "name",
                "id",
                "altitude"
              ],
              "type": "string",
              "description": "Field to sort by",
              "default": "date_modified",
              "title": "Sort By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Sort order",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of items to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Number of items to return"
          },
          {
            "name": "zenith_ra",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Zenith RA in degrees (required for altitude sort)",
              "title": "Zenith Ra"
            },
            "description": "Zenith RA in degrees (required for altitude sort)"
          },
          {
            "name": "zenith_dec",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Zenith Dec in degrees (required for altitude sort)",
              "title": "Zenith Dec"
            },
            "description": "Zenith Dec in degrees (required for altitude sort)"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date filter (YYYY-MM-DD format)",
              "title": "Date From"
            },
            "description": "Start date filter (YYYY-MM-DD format)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date filter (YYYY-MM-DD format)",
              "title": "Date To"
            },
            "description": "End date filter (YYYY-MM-DD format)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PfsDesignListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pfs_designs/positions": {
      "get": {
        "tags": [
          "pfs_designs",
          "pfs_designs"
        ],
        "summary": "Get PFS Design positions",
        "description": "Get positions (id, ra, dec) of PFS Designs. Supports filtering by search and date range.",
        "operationId": "list_design_positions_api_pfs_designs_positions_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search string (matches name or id)",
              "title": "Search"
            },
            "description": "Search string (matches name or id)"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date filter (YYYY-MM-DD format)",
              "title": "Date From"
            },
            "description": "Start date filter (YYYY-MM-DD format)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date filter (YYYY-MM-DD format)",
              "title": "Date To"
            },
            "description": "End date filter (YYYY-MM-DD format)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PfsDesignPosition"
                  },
                  "title": "Response List Design Positions Api Pfs Designs Positions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pfs_designs/{id_hex}.fits": {
      "get": {
        "tags": [
          "pfs_designs",
          "pfs_designs"
        ],
        "summary": "Download PFS Design FITS",
        "description": "Download the FITS file for a specific PFS Design.",
        "operationId": "download_design_api_pfs_designs__id_hex__fits_get",
        "parameters": [
          {
            "name": "id_hex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Hex"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pfs_designs/{id_hex}": {
      "get": {
        "tags": [
          "pfs_designs",
          "pfs_designs"
        ],
        "summary": "Get PFS Design details",
        "description": "Get detailed information about a specific PFS Design.",
        "operationId": "get_design_api_pfs_designs__id_hex__get",
        "parameters": [
          {
            "name": "id_hex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Hex"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PfsDesignDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PfsDesignDetail": {
        "properties": {
          "fits_meta": {
            "$ref": "#/components/schemas/FitsMeta"
          },
          "date_modified": {
            "type": "string",
            "format": "date-time",
            "title": "Date Modified"
          },
          "design_data": {
            "$ref": "#/components/schemas/DesignData"
          },
          "photometry_data": {
            "$ref": "#/components/schemas/PhotometryData"
          },
          "guidestar_data": {
            "$ref": "#/components/schemas/GuidestarData"
          }
        },
        "type": "object",
        "required": [
          "fits_meta",
          "date_modified",
          "design_data",
          "photometry_data",
          "guidestar_data"
        ],
        "title": "PfsDesignDetail",
        "description": "PFS Design \u8a73\u7d30"
      },
      "PfsDesignEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "frameid": {
            "type": "string",
            "title": "Frameid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date_modified": {
            "type": "string",
            "format": "date-time",
            "title": "Date Modified"
          },
          "ra": {
            "type": "number",
            "title": "Ra"
          },
          "dec": {
            "type": "number",
            "title": "Dec"
          },
          "arms": {
            "type": "string",
            "title": "Arms"
          },
          "num_design_rows": {
            "type": "integer",
            "title": "Num Design Rows"
          },
          "num_photometry_rows": {
            "type": "integer",
            "title": "Num Photometry Rows"
          },
          "num_guidestar_rows": {
            "type": "integer",
            "title": "Num Guidestar Rows"
          },
          "design_rows": {
            "$ref": "#/components/schemas/DesignRows"
          }
        },
        "type": "object",
        "required": [
          "id",
          "frameid",
          "name",
          "date_modified",
          "ra",
          "dec",
          "arms",
          "num_design_rows",
          "num_photometry_rows",
          "num_guidestar_rows",
          "design_rows"
        ],
        "title": "PfsDesignEntry",
        "description": "PFS Design \u30a8\u30f3\u30c8\u30ea\uff08\u4e00\u89a7\u7528\uff09"
      },
      "PfsDesignListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PfsDesignEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "PfsDesignListResponse",
        "description": "PFS Design \u4e00\u89a7\u30ec\u30b9\u30dd\u30f3\u30b9\uff08\u30da\u30fc\u30b8\u30cd\u30fc\u30b7\u30e7\u30f3\u5bfe\u5fdc\uff09"
      },
      "PfsDesignPosition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ra": {
            "type": "number",
            "title": "Ra"
          },
          "dec": {
            "type": "number",
            "title": "Dec"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ra",
          "dec"
        ],
        "title": "PfsDesignPosition",
        "description": "PFS Design \u4f4d\u7f6e\u60c5\u5831\uff08\u8efd\u91cf\u7248\uff09"
      },
...
    }
  }
}